\name{PPInfer-package}
\alias{PPInfer-package}
\alias{PPInfer}
\docType{package}
\title{
\packageTitle{PPInfer}
}
\description{
Interactions between proteins occur in many, if not most, biological processes. Most proteins perform their functions in networks associated with other proteins and other biomolecules. This fact has motivated the development of a variety of experimental methods for the identification of protein interactions. This variety has in turn ushered in the development of numerous different computational approaches for modeling and predicting protein interactions. Sometimes an experiment is aimed at identifying proteins closely related to some interesting proteins. A network based statistical learning method is used to infer the putative functions of proteins from the known functions of its neighboring proteins on a PPI network. This package identifies such proteins often involved in the same or similar biological functions.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PPInfer}
\packageIndices{PPInfer}

}
\author{ 
\packageAuthor{PPInfer}

Maintainer: \packageMaintainer{PPInfer}
}


